containerregistry package renamed to dockerkeychain
