+++ src/dockerkeychain/docker_keychain.go
@@ -1,4 +1,4 @@
-package containerregistry
+package dockerkeychain
 
 import (
 	"sync"
